+++ src/app/review/page.tsx
@@ -4,6 +4,7 @@ import { useState, useCallback } from 'react';
 
 import ImageUploader from '@components/common/ImageUploader';
 import StarRating from '@components/review/StarRating';
+import BottomNavigation from '@components/Home/BottomNavigation';
 
 export default function Page() {
   const [rating, setRating] = useState(0);
@@ -15,6 +16,7 @@ export default function Page() {
     <div>
       <StarRating rating={rating} onClick={handleRating} />
       <ImageUploader />
+      <BottomNavigation />
     </div>
   );
 }
