+++ src/components/Home/BottomNavigation/index.tsx
@@ -1,0 +1,18 @@
+import AddStoreButton from '../AddStoreButton';
+
+import CurrentLocation from '@components/Home/CurrentLocation';
+import ProfileButton from '@components/Home/ProfileButton';
+
+export default function BottomNavigation() {
+  const handleClickLocationButton = () => {
+    // NOTE: 현재 내 위치로 이동
+  };
+
+  return (
+    <div className="flex justify-between gap-[16px] w-full h-[128px] px-[24px] py-[16px]">
+      <CurrentLocation onClick={handleClickLocationButton} />
+      <AddStoreButton />
+      <ProfileButton />
+    </div>
+  );
+}
